Migrate _lemmas/_proof.go files to host-attached directives + inline helpers
Move all reqproof:lemma directives from
parser_delete_snippet_proof.go and parser_item3_lemmas_proof.go onto
the production functions they characterize (tokenEnd, tokenStart,
nextToken, lastToken, h2I) in parser.go and escape.go.

Inline the Delete-cleanup obligation helpers and the variadic
keysCount stand-in at the bottom of parser.go behind a
"// --- reqproof verification helpers ---" delimiter, since those
patterns abstract control-flow shapes that the translator does not
yet support directly on Delete.

Lemma corpus preserved: 30 total = 29 PROVED + 1 CE.
